Add user choice for dependency updates during installation of PCCS
When some known vulnerabilities in component dependencies are detected and session is interactive, user is asked what to do. Some automatic ways of updating dependencies are available for selection. If updating dependencies to newest versions is not enough (they are not fixed in newest public versions), user can select to continue with tested versions or abort the installation. In non-interactive mode package installation is cancelled if confirmation from the user is required.

---------

Signed-off-by: Marek Tomczewski <marek.tomczewski@intel.com>